+++ up-n-down/views/upload-form.php
@@ -18,7 +18,7 @@
 <?php foreach ( $this->files as $file ) : ?>
 	<li>
 		<a href="<?php echo $file['path']; ?>" title="Download <?php echo $file['name']; ?>" download="<?php echo $file['name']; ?>"><?php echo $file['name']; ?></a>
-		<?php if ( current_user_can( 'delete_pages' ) ) : ?>
+		<?php if ( current_user_can( 'delete_pages' ) || isset( $this->options['allow_subscribers_delete'] ) ) : ?>
 		<a href="<?php echo $this->permalink; ?>?upndown-delete-file=<?php echo $file['name']; ?>" title="Delete <?php echo $file['name']; ?>" class="upndown-delete-file">&times;</a>
 		<?php endif; ?>
 	</li>