+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Fix error with upload path.
 * Replace "require_once" with "include_once" for view files.
 * Fix global admin styles.
+* Allow subscribers to delete files as option.
 
 
 ## Version 1.0.0 ##